edms_logic/requirements.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
attrs==25.1.0
billiard==4.2.1
blessed==1.20.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
gunicorn==21.2.0
idna==3.6
importlib_resources==6.5.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jinxed==1.2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==23.1
prompt_toolkit==3.0.50
psycopg2==2.9.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2023.3
PyYAML==6.0.2
redis==3.5.3
referencing==0.36.2
requests==2.31.0
rpds-py==0.22.3
simplejson==3.20.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
